Flatten measurements in Observation
Flatten the temperature measurements in Observation. It is known
in advance how many measurements there are (two), and flattening
this object will make it convert cleaner to CSV. Add a
"temperatures" property which returns a tuple of channel
measurements, just like the old object did.

Flat is better than nested.